update: builtins and parseing merged
