+++ srcs/parse/tokenize_input.c
@@ -25,7 +25,6 @@ static int	init_cmds(t_mshell*mshell, char *input_str)
 		i++;
 	}
 	free(cmds_temp);
-	printf("\n");
 	return (0);
 }
 
@@ -64,7 +63,7 @@ int	tokenize_input(t_mshell *mshell, char *input_str)
 		// 	return (1);
 		i++;
 	}
-	print_command_list(mshell);
+	//print_command_list(mshell);
 	return (0);
 }
